Move decode function into it's own source file
